Updated
Back-end
Part 8. Ch04. 02.

+++ src/main/java/com/fastcampus/snsproject/controller/PostController.java
@@ -59,7 +59,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Integer postId, Authentication authentication) {
+    public Response<Long> likeCount(@PathVariable Integer postId, Authentication authentication) {
         return success(postService.likeCount(postId));
     }
